+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `JWT.createWithId` function
 
 ## 10.2.0 - 2026-04-14
 - Add `Session.EnforceScope.prepareWithPurpose` function